telethon/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
telethon==1.32.1
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
schedule==1.2.0
cryptography==41.0.7
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.2
qrcode[pil]>=7.4.2
websockets>=12.0
redis>=5.0.0
qdrant-client>=1.7.0
tiktoken>=0.5.0
apscheduler>=3.10.0
pytz>=2023.3
